-refactor reused code into utils
